renamed deatiledBar to detailedBar
