+++ src/components/charts/detailed-bar.js
